+++ dnssd/advertiser/internal/advertisertest/doc.go
@@ -1,0 +1,3 @@
+// Package advertisertest contains a standard test suite for [dnssd.Advertiser]
+// implementations.
+package advertisertest