+++ dnssd/advertiser/route53/doc.go
@@ -1,0 +1,3 @@
+// Package route53 provides a [dnssd.Advertiser] implementation that advertises
+// DNS-SD service instances on domain names hosted by Amazon Route 53.
+package route53